Dios que inteligente soy odio Linux y al padre de Linux y encima, a su abuela

+++ P1/FicherosP1/Exercises/Show_file/show_file.c
@@ -5,7 +5,7 @@
 int main(int argc, char* argv[]) {
 	FILE* file=NULL;
 	const int args = 3;
-	int c;
+	long int c;
 
 	if (argc != args) {
 		fprintf(stderr,"Usage: %s <file_name>\n",argv[0]);
